CATD-642 make AppGroupUserDefaultsStore private
