+++ RealifeTech/RealifeTech.swift
@@ -21,16 +21,10 @@ public class RealifeTech {
     public static var Sell: Sell!
     public static var CampaignAutomation: CampaignAutomation!
     public static var Identity: Identity!
-    public static var AppGroupStore: AppGroupUserDefaultsStore!
+    private static var AppGroupStore: AppGroupUserDefaultsStore!
 
     private static var moduleVersionString: String {
         Bundle(for: self.self).infoDictionary?["CFBundleShortVersionString"] as? String ?? "Unknown"
-    }
-
-    // TODO: refactor
-    private static var appGroupId: String {
-        let mainBundleId = Bundle.main.bundleIdentifier ?? "Unknown"
-        return "group." + mainBundleId
     }
 
     /// Provides information required for the SDK to operate.
@@ -79,8 +73,6 @@ public class RealifeTech {
                                     graphQLAPIUrl: configuration.graphQLApiUrl)
         Identity = IdentityFactory.makeModule(analyticsLogger: Analytics,
                                               identityPersister: identityPersister)
-        AppGroupStore = AppGroupUserDefaultsStore(appGroupId: appGroupId)
-        AppGroupStore.saveSDKConfiguration(with: configuration)
     }
 
     private static func createAPIHelper(with configuration: SDKConfiguration, deviceId: String) -> APITokenManagable {
@@ -120,5 +112,10 @@ public class RealifeTech {
     /// - Parameter webOrderingJourneyUrl: New URL with the string type
     public static func set(webOrderingJourneyUrl: String) {
         Sell.orderingJourneyUrl = webOrderingJourneyUrl
+    }
+
+    public static func setNotificationServiceExtensionWith(appGroupId: String, configuration: SDKConfiguration) {
+        AppGroupStore = AppGroupUserDefaultsStore(appGroupId: appGroupId)
+        AppGroupStore.saveSDKConfiguration(with: configuration)
     }
 }
