+++ SampleApp/NotificationServiceExtension/NotificationService.swift
@@ -14,23 +14,14 @@ class NotificationService: UNNotificationServiceExtension {
 
     private var contentHandler: ((UNNotificationContent) -> Void)?
     private var bestAttemptContent: UNMutableNotificationContent?
+    private let configurator = RLTNotificationServiceConfigurator(appGroupId: "group.com.concertlive.SampleApp")
 
     override func didReceive(_ request: UNNotificationRequest,
                              withContentHandler contentHandler: @escaping (UNNotificationContent) -> Void) {
         self.contentHandler = contentHandler
         self.bestAttemptContent = (request.content.mutableCopy() as? UNMutableNotificationContent)
         os_log("[DEBUG] ---NotificationService didReceive", log: .default, type: .error)
-        let store = AppGroupUserDefaultsStore(appGroupId: "group.com.concertlive.SampleApp")
-        RLTNotificationServiceConfigurator(appGroupStore: store).didReceive(request, withContentHandler: contentHandler)
-
-//        if let bestAttemptContent = bestAttemptContent,
-//           let userInfo = bestAttemptContent.userInfo as? [String: Any] {
-//            configurator.fetchAppSecretAndConfigureSDK {
-//                RealifeTech.Communicate.trackPush(event: .received, trackInfo: userInfo) { _ in
-//                    contentHandler(bestAttemptContent)
-//                }
-//            }
-//        }
+        configurator.didReceive(request, withContentHandler: contentHandler)
     }
 
     override func serviceExtensionTimeWillExpire() {
